backend/requirements.txt
# FastAPI
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
python-dotenv==1.1.0
starlette==0.46.1

# Database
SQLAlchemy==2.0.41

# LLM
openai==1.70.0

# ASR
faster-whisper==1.1.1

# Video processing
ffmpeg-python==0.2.0
pillow==11.0.0
numpy==2.2.4

# Document generation
docxtpl==0.19.0
python-docx==1.2.0

# Utils
pydantic==2.11.2
requests==2.32.3
PyYAML==6.0.2
aiofiles==24.1.0
